Fix beyond-i64 collection typing for C#, Dart, and VB
Require beyond_i64_type at opener call sites, widen C#/Dart/VB
homogeneous containers past signed 64-bit, and quiet pylint spelling
on the new technical terms.

+++ tests/integration/cases/list_beyond_i64/VisualBasic@dotnet_6.vb
@@ -1,6 +1,6 @@
 Imports System.Collections.Generic
 Module Check
-    Dim my_data = New Integer() {
+    Dim my_data = New ULong() {
         9223372036854775807UL,
         9223372036854775808UL
     }
