+++ tests/integration/cases/list_beyond_i64/VisualBasic_combined@dotnet_6.vb
@@ -1,14 +1,14 @@
 Imports System.Collections.Generic
 Module Check
     Sub _declaration()
-        Dim my_data = New Integer() {
+        Dim my_data = New ULong() {
             9223372036854775807UL,
             9223372036854775808UL
         }
     End Sub
     Sub _assignment()
         Dim my_data As Object
-        my_data = New Integer() {
+        my_data = New ULong() {
             9223372036854775807UL,
             9223372036854775808UL
         }
