+++ tests/integration/cases/map_beyond_i64/CSharp@v10.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-var my_data = new Dictionary<string, int> {
-    ["a"] = 9223372036854775807,
-    ["b"] = 9223372036854775808
+var my_data = new Dictionary<string, ulong> {
+    ["a"] = 9223372036854775807UL,
+    ["b"] = 9223372036854775808UL
 };
